chore: move to separate 'package'
