+++ fontless/vite.ts
@@ -11,7 +11,8 @@ import type { FontlessOptions } from './types'
 import { resolveMinifyCssEsbuildOptions, transformCSS } from './utils'
 import type { FontFamilyInjectionPluginOptions } from './utils'
 import { createResolver } from './resolve'
-import { normalizeFontData, type NormalizeFontDataContext } from './assets'
+import { normalizeFontData } from './assets'
+import type { NormalizeFontDataContext } from './assets'
 import { storage } from './storage'
 import { resolveProviders } from './providers'
