+++ package.json
@@ -54,6 +54,7 @@
     "defu": "^6.1.4",
     "esbuild": "^0.25.4",
     "fontaine": "^0.6.0",
+    "fontless": "file:./fontless",
     "h3": "^1.15.3",
     "jiti": "^2.4.2",
     "magic-regexp": "^0.10.0",
